backend/requirements.txt
fastapi
uvicorn
requests
numpy
scipy
pydantic
python-dotenv
google-generativeai
geopy